Move currency generation code to a custom hook
